Fix SOpenAdapterArgs struct in DDI
Since the original PrivateCallbacks struct was just embedded in the
middle of the SOpenAdapterArgs struct (not a pointer), adding new fields
breaks compatibility between version, yikes! This commit adds a new
PrivateCallbacks2 struct that is referenced with a pointer at the end of
the SOpenAdapterArgs struct and includes the new present callback to
make things compatible again.